Merged PR 282: Added clickable links
The new component allows for a lot of extensibility. I could've added this to all texts, and rendered accordingly to a "allowLinks" attribute.

Since we just need links in messages, I went with the approach of having it as a separate component. This prevents unnecessary checks everywhere else.

We'll have to use this component uses `NumberlessText` under the hood, which makes it customisable either way.

- added autolinking to text
- text update
- added click links to replies and text