Merged PR 838: Retrying sent reactions while offline
Test Cases: React to a message while in airplane mode, then turn on the internet and see if it gets delivered to your other device.

Earlier, we were not retrying sending the reaction once the user came back online because the message was not journaled. In this PR,
- I am changing the message status to journaled and saving it in DB and
- we will call the retry method. It will reattempt sending the message/reactions with the previously saved payload.
- if the sendObject in retry fails or gives un status send, then we will run cleanup to clean delete message.

_Check the implementation of the generic sender retry method; it is quite similar except for the database saving part._

Related work items: #694