Merged PR 937: Contact sharing flow if user has permissions off
This pr contains contact sharing flow specifically if the user permissions are off.
This pr has changes in UI as well as functionality.
2 addtional bubbles have been added:
ContactInfoBubble
ContactRequestBubble
Flow:
If the Alice's permissions are off and Bob wants to send her contact to Charlie,
Bob will send a contact request to Alice. This will check if contact permissions are on.
If they are on, a port bundle will be generated and sent over to Bob who will then send it to Bob.
If they are off, Alice will have an option to allowing once which will generate a bundle and send it to Bob who will then send it to Charlie.

Appropriate doc strings have been mentioned in utils.

link to designs: https://www.figma.com/design/AkuLAhxtmpIujnwFExvdJe/Port-Final_V_01?node-id=16154-66270&t=xf8DZ4nTqqX6Peum-4