Merged PR 1072: Returned the notification permission on iOS
The notification permission was hidden on iOS due to considerations regarding notification filtering. This, along with a corresponding back end PR adds support for server side silencing of notifications.

It achieves silencing by making an API call when a chat's permissions for notifications are updated. Attempts have been made to minimize the number or network calls made.

It has considered 3 primary cases:
- Editing notification permissions for an existing chat
- Applying notification settings to existing contacts in a folder
- Creating a new chat and inheriting notification permissions from the folder it is dropped into