Merged PR 1089: Templates on chat screen
This pr contains the feature for templates on chat screen.
In popupactions, you will have a new button for templates.
A user can create a template, edit and delete a template.
Once a template is clicked, the message bar automatically gets populated.
The user can then edit the message as well.