Merged PR 1400: Force close all modals if necessary
Some modals are not screen specific and can open up due to certain triggers and can conflict with modals that are already open.
For this, we now have a redux store call to force close all existing modals.
This call can be used before opening a modal that's not screen specific.