Remove background message pull backlog on Android (#274)
On Android, we're getting reports of message latency, particularly in the killed state. This could be because we're using the network when we have been given very limited amounts of background processing time. To combat this, we attempt to simply process the sent payload in the background state. In the foreground state we can pullBacklog as usual since we're afforded more processing time.